Fix issues 3 (segfault on ARM) & 4 (misdisassembly of ARM) They correspond to binutils issue
 http://sourceware.org/bugzilla/show_bug.cgi?id=13135

pulled in Stephen's ARM fix, and added a MIPS compile fix (could
do with a check by a MIPS head).

Dave

+++ binutils/binutils-2.22/opcodes/arm-dis.c
@@ -1891,7 +1891,7 @@ print_insn_coprocessor (bfd_vma pc,
 		      {
 			if (offset)
 			  func (stream, ", #%d]%s",
-				offset,
+				(int)offset,
 				WRITEBACK_BIT_SET ? "!" : "");
 			else if (NEGATIVE_BIT_SET)
 			  func (stream, ", #-0]");
@@ -1905,7 +1905,7 @@ print_insn_coprocessor (bfd_vma pc,
 			if (WRITEBACK_BIT_SET)
 			  {
 			    if (offset)
-			      func (stream, ", #%d", offset);
+			      func (stream, ", #%d", (int)offset);
 			    else if (NEGATIVE_BIT_SET)
 			      func (stream, ", #-0");
 			  }
@@ -1913,7 +1913,7 @@ print_insn_coprocessor (bfd_vma pc,
 			  {
 			    func (stream, ", {%s%d}",
 				  (NEGATIVE_BIT_SET && !offset) ? "-" : "",
-				  offset);
+				  (int)offset);
 			    value_in_comment = offset;
 			  }
 		      }
@@ -2224,7 +2224,7 @@ print_insn_coprocessor (bfd_vma pc,
 		    {
 		      /* given (20, 23) | given (0, 3) */
 		      value = ((given >> 16) & 0xf0) | (given & 0xf);
-		      func (stream, "%d", value);
+		      func (stream, "%d", (int)value);
 		    }
 		    break;
 
@@ -2349,7 +2349,7 @@ print_arm_address (bfd_vma pc, struct disassemble_info *info, long given)
 	  /* Pre-indexed.  Elide offset of positive zero when
 	     non-writeback.  */
 	  if (WRITEBACK_BIT_SET || NEGATIVE_BIT_SET || offset)
-	    func (stream, ", #%s%d", NEGATIVE_BIT_SET ? "-" : "", offset);
+	    func (stream, ", #%s%d", NEGATIVE_BIT_SET ? "-" : "", (int)offset);
 
 	  if (NEGATIVE_BIT_SET)
 	    offset = -offset;
@@ -2364,7 +2364,7 @@ print_arm_address (bfd_vma pc, struct disassemble_info *info, long given)
 	}
       else  /* Post indexed.  */
 	{
-	  func (stream, "], #%s%d", NEGATIVE_BIT_SET ? "-" : "", offset);
+	  func (stream, "], #%s%d", NEGATIVE_BIT_SET ? "-" : "", (int)offset);
 
 	  /* Ie ignore the offset.  */
 	  offset = pc + 8;
@@ -2386,7 +2386,7 @@ print_arm_address (bfd_vma pc, struct disassemble_info *info, long given)
 	      /* Elide offset of positive zero when non-writeback.  */
 	      offset = given & 0xfff;
 	      if (WRITEBACK_BIT_SET || NEGATIVE_BIT_SET || offset)
-		func (stream, ", #%s%d", NEGATIVE_BIT_SET ? "-" : "", offset);
+		func (stream, ", #%s%d", NEGATIVE_BIT_SET ? "-" : "", (int)offset);
 	    }
 	  else
 	    {
@@ -2404,7 +2404,7 @@ print_arm_address (bfd_vma pc, struct disassemble_info *info, long given)
 	      /* Always show offset.  */
 	      offset = given & 0xfff;
 	      func (stream, "], #%s%d",
-		    NEGATIVE_BIT_SET ? "-" : "", offset);
+		    NEGATIVE_BIT_SET ? "-" : "", (int)offset);
 	    }
 	  else
 	    {
@@ -3002,7 +3002,7 @@ print_insn_arm (bfd_vma pc, struct disassemble_info *info, long given)
 			      /* Elide positive zero offset.  */
 			      if (offset || NEGATIVE_BIT_SET)
 				func (stream, "[pc, #%s%d]\t; ",
-				      NEGATIVE_BIT_SET ? "-" : "", offset);
+				      NEGATIVE_BIT_SET ? "-" : "", (int)offset);
 			      else
 				func (stream, "[pc]\t; ");
 			      if (NEGATIVE_BIT_SET)
@@ -3013,7 +3013,7 @@ print_insn_arm (bfd_vma pc, struct disassemble_info *info, long given)
 			    {
 			      /* Always show the offset.  */
 			      func (stream, "[pc], #%s%d",
-				    NEGATIVE_BIT_SET ? "-" : "", offset);
+				    NEGATIVE_BIT_SET ? "-" : "", (int)offset);
 			      if (! allow_unpredictable)
 				is_unpredictable = TRUE;
 			    }
@@ -3182,7 +3182,7 @@ print_insn_arm (bfd_vma pc, struct disassemble_info *info, long given)
 			  {
 			    if (offset)
 			      func (stream, ", #%d]%s",
-				    value_in_comment,
+				    (int)value_in_comment,
 				    WRITEBACK_BIT_SET ? "!" : "");
 			    else
 			      func (stream, "]");
@@ -3194,11 +3194,11 @@ print_insn_arm (bfd_vma pc, struct disassemble_info *info, long given)
 			    if (WRITEBACK_BIT_SET)
 			      {
 				if (offset)
-				  func (stream, ", #%d", value_in_comment);
+				  func (stream, ", #%d", (int)value_in_comment);
 			      }
 			    else
 			      {
-				func (stream, ", {%d}", offset);
+				func (stream, ", {%d}", (int)offset);
 				value_in_comment = offset;
 			      }
 			  }
@@ -3240,7 +3240,7 @@ print_insn_arm (bfd_vma pc, struct disassemble_info *info, long given)
 			  if (name != NULL)
 			    func (stream, "%s", name);
 			  else
-			    func (stream, "(UNDEF: %lu)", sysm);
+			    func (stream, "(UNDEF: %lu)", (unsigned long)sysm);
 			}
 		      else
 			{
@@ -3404,7 +3404,7 @@ print_insn_arm (bfd_vma pc, struct disassemble_info *info, long given)
 			  if (name != NULL)
 			    func (stream, "%s", name);
 			  else
-			    func (stream, "(UNDEF: %lu)", sysm);
+			    func (stream, "(UNDEF: %lu)", (unsigned long)sysm);
 			}
 			break;
 
@@ -3636,17 +3636,17 @@ print_insn_thumb16 (bfd_vma pc, struct disassemble_info *info, long given)
 			    break;
 
 			  case 'd':
-			    func (stream, "%ld", reg);
+			    func (stream, "%ld", (long)reg);
 			    value_in_comment = reg;
 			    break;
 
 			  case 'H':
-			    func (stream, "%ld", reg << 1);
+			    func (stream, "%ld", (long)(reg << 1));
 			    value_in_comment = reg << 1;
 			    break;
 
 			  case 'W':
-			    func (stream, "%ld", reg << 2);
+			    func (stream, "%ld", (long)(reg << 2));
 			    value_in_comment = reg << 2;
 			    break;
 
@@ -3660,7 +3660,7 @@ print_insn_thumb16 (bfd_vma pc, struct disassemble_info *info, long given)
 			    break;
 
 			  case 'x':
-			    func (stream, "0x%04lx", reg);
+			    func (stream, "0x%04lx", (long)reg);
 			    break;
 
 			  case 'B':
@@ -3966,11 +3966,11 @@ print_insn_thumb32 (bfd_vma pc, struct disassemble_info *info, long given)
 		    }
 
 		  if (postind)
-		    func (stream, "], #%d", offset);
+		    func (stream, "], #%d", (int)offset);
 		  else
 		    {
 		      if (offset)
-			func (stream, ", #%d", offset);
+			func (stream, ", #%d", (int)offset);
 		      func (stream, writeback ? "]!" : "]");
 		    }
 
@@ -4199,7 +4199,7 @@ print_insn_thumb32 (bfd_vma pc, struct disassemble_info *info, long given)
 		    if (name != NULL)
 		      func (stream, "%s", name);
 		    else
-		      func (stream, "(UNDEF: %lu)", sysm);
+		      func (stream, "(UNDEF: %lu)", (unsigned long)sysm);
 		  }
 		else
 		  {
@@ -4221,7 +4221,7 @@ print_insn_thumb32 (bfd_vma pc, struct disassemble_info *info, long given)
 		    if (name != NULL)
 		      func (stream, "%s", name);
 		    else
-		      func (stream, "(UNDEF: %lu)", sm);
+		      func (stream, "(UNDEF: %lu)", (unsigned long)sm);
 		  }
 		else
 		  func (stream, psr_name (given & 0xff));
